1.add Exception 2.add LibraryFile 3.modify for unit test

+++ io/FileInputStream.cpp
@@ -16,7 +16,7 @@ int _FileInputStream::read() {
     return 0;
 }
 
-int _FileInputStream::read(ByteArray buff) {
+long _FileInputStream::read(ByteArray buff) {
     fstream.read(buff->toValue(),buff->size());
     return fstream.gcount();
 }
