+++ io/include/ZipStream.hpp
@@ -32,7 +32,7 @@ public:
 
     int read();
 
-    int read(ByteArray buffer);
+    long read(ByteArray buffer);
 
     bool open();
